pyproject.toml
[project]
name = "diabot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "fastembed>=0.7.1",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.4",
    "pandas>=2.3.1",
    "python-calamine>=0.4.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "uvicorn>=0.35.0",
]
